IGNITE-9026 fix random class loading failures
Skip recursive resource requests to orginating nodes, rather than failing the entire request.   Continue to search other nodes on errors, because assumption that all nodes have the same view is incorrect.
Restrict the recursive searches that a node should do when looking for resources by avoiding the nodes that the sender has or will search.